requirements.txt
beautifulsoup4==4.9.3
clean-text==0.4.0
lxml==4.6.3
pymongo==3.11.3
python-dotenv==0.16.0
requests==2.25.1
Unidecode==1.2.0
xmltodict==0.12.0
